Add files via upload
I worked on the GUI, got the help and about buttons working but the text is just quickly written for about and help is just getting started from our user guide copy and pasted. Add now puts whatever is in the text fields into a row. When a row is selected, it's data reappears in the text fields, and if you change them and press edit it will change the row. All of the data from the text field is string though so that will need to be changed.